Seperate server executeable and api controllers
