+++ BeatTogether.Status.Api.Controllers/Configuration/QuickplayConfiguration.cs
@@ -1,0 +1,11 @@
+﻿using BeatTogether.Status.Api.Controllers.Models;
+using System.Collections.Generic;
+
+namespace BeatTogether.Status.Api.Controllers.Configuration
+{
+    public class QuickplayConfiguration
+    {
+        public List<PredefinedPack> PredefinedPacks { get; set; } = new();
+        public List<LocalizedCustomPack> LocalizedCustomPacks { get; set; } = new();
+    }
+}
