+++ BeatTogether.Status.Api.Controllers/Models/LocalizedCustomPackName.cs
@@ -1,0 +1,10 @@
+﻿using BeatTogether.Status.Api.Controllers.Enums;
+
+namespace BeatTogether.Status.Api.Controllers.Models
+{
+    public class LocalizedCustomPackName
+    {
+        public Language language { get; set; }
+        public string packName { get; set; } = string.Empty;
+    }
+}
