+++ BeatTogether.Status.Api.Controllers/Models/MasterServerQuickplayData.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace BeatTogether.Status.Api.Models
+namespace BeatTogether.Status.Api.Controllers.Models
 {
     public record MasterServerQuickplayData(QuickplaySongPacksOverride quickPlayAvailablePacksOverride);
 }
