+++ BeatTogether.Status.Api.Controllers/Models/MasterServerStatusData.cs
@@ -1,7 +1,7 @@
-﻿using BeatTogether.Status.Api.Enums;
+﻿using BeatTogether.Status.Api.Controllers.Enums;
 using System.Collections.Generic;
 
-namespace BeatTogether.Status.Api.Models
+namespace BeatTogether.Status.Api.Controllers.Models
 {
     public record MasterServerStatusData(
         string minimumAppVersion,
