+++ BeatTogether.Status.Api.Controllers/Models/PredefinedPack.cs
@@ -1,0 +1,8 @@
+﻿namespace BeatTogether.Status.Api.Controllers.Models
+{
+    public class PredefinedPack
+    {
+        public int order { get; set; }
+        public string packId { get; set; } = string.Empty;
+    }
+}
