+++ BeatTogether.Status.Api.Controllers/Models/UserMessage.cs
@@ -1,0 +1,5 @@
+﻿
+namespace BeatTogether.Status.Api.Controllers.Models
+{
+    public record UserMessage(List<LocalizedMessage> localizedMessages);
+}
