+++ BeatTogether.Status.Api/.dockerignore
