+++ BeatTogether.Status.Api/BeatTogether.Status.Api.csproj
@@ -1,0 +1,13 @@
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <LangVersion>latest</LangVersion>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\BeatTogether.Status.Api.Controllers\BeatTogether.Status.Api.Controllers.csproj" />
+  </ItemGroup>
+
+</Project>
