+++ BeatTogether.Status.Api/BeatTogether.Status.Api.sln
@@ -1,9 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30626.31
+# Visual Studio Version 17
+VisualStudioVersion = 17.8.34316.72
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BeatTogether.Status.Api", "BeatTogether.Status.Api.csproj", "{EF1C2D6B-3B1A-48A6-BD7F-554910E7E957}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BeatTogether.Status.Api", "BeatTogether.Status.Api.csproj", "{EF1C2D6B-3B1A-48A6-BD7F-554910E7E957}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BeatTogether.Status.Api.Controllers", "..\BeatTogether.Status.Api.Controllers\BeatTogether.Status.Api.Controllers.csproj", "{5E167F5A-625A-4F74-8424-FE6FD428D904}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{EF1C2D6B-3B1A-48A6-BD7F-554910E7E957}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EF1C2D6B-3B1A-48A6-BD7F-554910E7E957}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EF1C2D6B-3B1A-48A6-BD7F-554910E7E957}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5E167F5A-625A-4F74-8424-FE6FD428D904}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5E167F5A-625A-4F74-8424-FE6FD428D904}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5E167F5A-625A-4F74-8424-FE6FD428D904}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5E167F5A-625A-4F74-8424-FE6FD428D904}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
