+++ BeatTogether.Status.Api/Dockerfile
