+++ BeatTogether.Status.Api/Program.cs
@@ -1,5 +1,5 @@
 using BeatTogether.Extensions;
-using BeatTogether.Status.Api.Configuration;
+using BeatTogether.Status.Api.Controllers.Configuration;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
@@ -40,17 +40,5 @@ namespace BeatTogether.Status.Api
                         )
                 )
                 .UseSerilog();
-    }
-
-    public static class HostBuilderExtensions
-    {
-        public static IHostBuilder UseStatusServer(this IHostBuilder hostBuilder) =>
-            hostBuilder
-                .ConfigureAppConfiguration()
-                .ConfigureServices((hostBuilderContext, services) =>
-                    services
-                        .Configure<StatusConfiguration>(hostBuilderContext.Configuration.GetSection("Status"))
-                        .Configure<QuickplayConfiguration>(hostBuilderContext.Configuration.GetSection("Quickplay"))
-                );
     }
 }
