+++ BeatTogether.Status.Api/Properties/launchSettings.json
