+++ BeatTogether.Status.Api/appsettings.json
