--- Models/PredefinedPack.cs
@@ -1,8 +1,0 @@
-﻿namespace BeatTogether.Status.Api.Models
-{
-    public class PredefinedPack
-    {
-        public int order { get; set; }
-        public string packId { get; set; } = string.Empty;
-    }
-}
