--- Models/UserMessage.cs
@@ -1,6 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace BeatTogether.Status.Api.Models
-{
-    public record UserMessage(List<LocalizedMessage> localizedMessages);
-}
